cast to uint32_t instead of uint16_t for receive on teensy

+++ src/teensy_can.cpp
@@ -91,7 +91,7 @@ _MB_ptr TeensyCAN<bus_num>::ProcessMessage = [](const CAN_message_t &msg)
 {
     std::array<uint8_t, 8> msg_data{};
     memcpy(msg_data.data(), msg.buf, 8);
-    CANMessage received_message{static_cast<uint16_t>(msg.id), msg.len, msg_data};
+    CANMessage received_message{static_cast<uint32_t>(msg.id), msg.len, msg_data};
     for (size_t i = 0; i < rx_messages_.size(); i++)
     {
         if (rx_messages_[i]->GetID() == received_message.id_)
